port CrashRpt to x64: * add platform to vcproj file * update TortoiseSVN.sln accordingly * made usage of DbgHelp API platform neutral   (mainly, use DWORD_PTR instead of DWORD) * remove linker pragmas from CrashRpt.cpp (export "C" does the same)   as they broke linkage on x64
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@6616 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/crashrpt/MailMsg.cpp
@@ -156,11 +156,11 @@ int CMailMsg::MAPISend()
 
    TStrStrVector::iterator p;
    int                  nIndex = 0;
-   int                  nRecipients = 0;
+   size_t               nRecipients = 0;
    MapiRecipDesc*       pRecipients = NULL;
    MapiRecipDesc*       pOriginator = NULL;
    MapiRecipDesc*       pFirstRecipient = NULL;
-   int                  nAttachments = 0;
+   size_t               nAttachments = 0;
    MapiFileDesc*        pAttachments = NULL;
    ULONG                status = 0;
    MapiMessage          message;
